StaticContentPlugin: Actually serve real files
For now, assume all files are text/html.

Also, tell the client the full server path to any
files that could not be found, opened, or read.
This isn't a smart thing to do in production, but
we'll let it slide for now since we're still testing
and developing the server.