ChatRoomPlugin: fix build sequence issue
If the plug-in was changed, but the tests for it weren't,
the tests were ending up testing an old version of the plug-in.

To prevent this from happening, rather that having a post-build
step for the tests that made a copy of the plug-in, just build
the plug-in and its tests into the same output folder.